Converted array element counter to long to avoid overflow

+++ hollow/src/main/java/com/netflix/hollow/core/write/HollowListTypeWriteState.java
@@ -160,7 +160,7 @@ public class HollowListTypeWriteState extends HollowTypeWriteState {
 
         ByteData data = ordinalMap.getByteData().getUnderlyingArray();
 
-        int elementCounter[] = new int[numShards];
+        long elementCounter[] = new long[numShards];
         int shardMask = numShards - 1;
 
         for(int ordinal=0;ordinal<=maxOrdinal;ordinal++) {
